# lib/api-zod/src/generated/types/cancelOrderRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BSV DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */

export interface CancelOrderRequest {
  walletAddress: string;
  signedTx?: string;
}