# lib/api-zod/src/generated/types/market.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BSV DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { MarketStatus } from "./marketStatus";
import type { MarketType } from "./marketType";

export interface Market {
  symbol: string;
  baseAsset: string;
  quoteAsset: string;
  lastPrice: number;
  priceChange24h: number;
  priceChangePercent24h: number;
  volume24h: number;
  high24h: number;
  low24h: number;
  marketCap?: number;
  status: MarketStatus;
  type: MarketType;
  minOrderSize?: number;
  maxOrderSize?: number;
  tickSize?: number;
  makerFee: number;
  takerFee: number;
}