# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BSV DEX Exchange API
 * OpenAPI spec version: 0.1.0
 */
import type { OrderSide } from "./orderSide";
import type { OrderStatus } from "./orderStatus";
import type { OrderTimeInForce } from "./orderTimeInForce";
import type { OrderType } from "./orderType";

export interface Order {
  id: string;
  symbol: string;
  walletAddress: string;
  side: OrderSide;
  type: OrderType;
  status: OrderStatus;
  price?: number;
  stopPrice?: number;
  quantity: number;
  filledQuantity: number;
  remainingQuantity: number;
  total?: number;
  fee: number;
  feeAsset: string;
  timeInForce: OrderTimeInForce;
  createdAt: Date;
  updatedAt: Date;
  txid?: string;
  /** BSV signed transaction hex */
  signedTx?: string;
}